# openapi.yaml
openapi: 3.1.0
info:
  title: Mock Orders API
  description: API mock per recuperare lo stato degli ordini clienti
  version: 1.0.0
servers:
  - url: https://your-domain-or-localtunnel/api
paths:
  /orders/{id}:
    get:
      operationId: getOrderById
      summary: Recupera i dettagli di un ordine tramite ID
      parameters:
        - name: id
          in: path
          required: true
          description: ID univoco dell'ordine
          schema:
            type: string
      responses:
        "200":
          description: Dati dell'ordine trovati
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Ordine non trovato
  /orders/search:
    get:
      operationId: searchOrdersByEmail
      summary: Cerca ordini per email del cliente
      description: Restituisce tutti gli ordini associati a un indirizzo email specifico.
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
            format: email
          description: Indirizzo email del cliente da cercare
      responses:
        "200":
          description: Lista di ordini trovati
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items:
                      $ref: "#/components/schemas/Order"
        "400":
          description: Parametro email mancante o non valido
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        "404":
          description: Nessun ordine trovato per l'email fornita
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

components:
  schemas:
    Order:
      type: object
      properties:
        id:
          type: string
        customer:
          type: string
        customerEmail:
          type: string
        status:
          type: string
        date:
          type: string
        total:
          type: number
        items:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              qty:
                type: integer
              price:
                type: number